dev-requirements-py2.txt
behave==1.2.6
behaving==2.0.0
Appium-Python-Client<=0.52
ckanapi==4.3
ckantoolkit>=0.0.4
factory-boy==2.1.1
pytest-factoryboy==1.3.2
flake8==3.8.3
mock
pytest-ckan
six>=1.13.0
splinter>=0.13.0,<0.17
xlrd==1.2.0
python-magic==0.4.18
messytables==0.15.2
progressbar==2.5

pika>=1.1.0,<1.3.0 # resolve harvester conflict
